Suggested edits around references to AIS and DBP
Offering a suggestion for consistent use of a short name for Alfresco Identity Service.  A short name of 'Identity Service' was used and if wanting to do this perhaps it is best to clear the name by adding it in parenthesis after the first reference to the Alfresco Identity Service (Identity Service) or optionally what I've proposed in this PR is to go with AIS.  Further and similar I noticed the use of DBP without first clearing it in the document which I've proposed in the first use of the Digital Business Platform (DBP).